Convert books function to module
